requirements.txt
asgiref==3.8.1
certifi==2025.1.31
cffi==1.17.1
channels==4.2.0
charset-normalizer==3.4.1
cryptography==44.0.0
Django==5.1.5
djangorestframework==3.15.2
idna==3.10
pycparser==2.22
requests==2.32.3
sqlparse==0.5.3
urllib3==2.3.0
